Initial commit of validation using LinkML
These replacement functions duplicate behaviour, so, at least for now,
these are called things like `new_validate'. Naming these is quite hard
and coming up with better names for these + usage instructions can be
done later.

The job manifest is now serialised as RDF, but using the JSON-LD
dumper/loader should be a drop-in replacement, if we wanted to use
JSON-LD instead of RDF for these manifests.

The test `job.yaml' schema which provides @context to JSON-LD and RDF
job manifests is stored in a different repository for now (the path
to `examples/linkml-scratch/working/' is just a placeholder as I stored
my work there during development).

I will commit the initial data model + YAML schemata examples after
making the CLI behave as before. Documentation of how I used the LinkML
API for this (referencing both the command-line tools + API part), and
the various 'quirks' I encountered should be productive.

I have used the `PurePath' class from `pathlib' in many of these
functions. This arises from issues with GCP when relative directories
like ".." are provided: it creates a sub-folder called ".." in the
bucket sub-directory in question, and files in that directory 404
until they are moved to the parent directory. The `PurePath' class is
helpful because it can extract the path base-name, file extension
suffixes &c. However, I haven't actually tested this yet.

Finally, I have followed LinkML in using type hinting for these
functions. This is very helpful and a boon for correctness and
structuring of the code.