Add example data with underscores instead of dots in column names
The underscores issue is simply for any conversion of the schema and
data as a single file with JSON-LD context or similar.

Validation does not depend on the slots having no dots in names.

This is primarily for purposes of demonstration (see previous long
commit message).

This further demonstrates treatment of missing data without the
[non-functional] `NA_None' type.

The schema and data files with underscores instead of dots have the
suffix `_us'.