Fully convert to LinkML and add verbosity/dry-run flags
The `fisdat' and `fisup' commands now support three levels of verbosity:
1. Normal messages (only really errors like a validation error)
2. Messages which reveal specific operations
3. Messages which reveal the function being run, and its arguments

To enable this, I added two tiny helper functions which are always fed
the verbosity level. `vprint()' will print a message if verbosity is set
to either level 2/3, while `vvprint()' will print a message if verbosity
is set to level 3 only.

The two commands now look for the data model relative to their absolute
path, so running these in the `examples/' directory now works.

As regards to `fisup', the CLI now works exactly like before (with the
addition of verbosity and dry-run options), so remove the old CLI
entirely. It was primarily there as a reference.