Include version number in string and move to pyproject.toml
Semantic versioning like this is non-trivial. The way to include
information about a given commit in the program distribution seems to be
to use something like `versioneer' or `bumpversion':

https://stackoverflow.com/questions/51071891/auto-increment-version-number-in-a-python-webserver-with-git/51076605#51076605

There also exists `setuptools.scm', but this does not seem to be as
readily accessible as versioneer.

This commit uses the `versioneer' to print the version number in the
program whenever it is run. This will reveal if we're running the most
recent version, or what version is otherwise being run.

The side effect of this is instead of setuptools to use pip directly,
with a shim `setup.py' and `pyproject.toml' instead. The equivalent of
`setup.py develop' is `pip install -e .', which seems to be more
predictable.

The first part of the version number seems to be derived from `git-tag'.