Make sure to serialise jobs section upon manifest initialisation
This makes sure that these fields, even if empty, are filled out.

This is in preparation for being able to write job specifications
as a YAML template. In both cases, we treat the manifest as *data*
with the `job.yaml' component of the data model as its *schema*, in
addition to LinkML SchemaView object / JSON-LD context.

The idea is to be able to convert a turtle manifest to the equivalent
YAML representation, in which we fill out the job specification.

Consequently, I have added a command-line flag to `fisdat' called
`--example-job'. When this is unset, upon initialisation of the
manifest (rather than appending), an almost completely-empty job
section is added. When this is set, the fields get filled out.
In both cases, the job type is set to "ignore" to signal that the job
in this section should never be executed.

Appending to the manifest assumes that this section exists already.
The data model's `job.yaml' section will be updated accordingly to
require this. This avoids writing a number of duplicate empty jobs
to that section.