Improve manifest/template conversion CLI
This now has generic arguments, mode, input and output. Previously, the
position of YAML template and RDF/TTL manifest was the same, regardless
of mode, which was not intuitive at all.

Furthermore, don't overwrite files by default. The previous behaviour was
fairly annoying, and while the `fisdat' program also works on existing
files, it is appending to them. So, previous behaviour of `fisjob' was
not consistent with `fisdat'.