Set identifier for other classes in `job.yaml' schema component
This was non-trivial, as `job_sources' (class `SourceDesc') seemed to
get coerced to a string when embedded in the example table description.

I am not certain of the mechanism behind this, but setting it to an
empty list and then using the `.append()' method of the class instance
seemed to fix this.

I further suspect a bug in the RDF generation, going by some of the
logs.

To enable this conversion to use identifiers in the YAML `job.yaml'
schema (and thus name the parts in the generated RDF), I had to
make these attributes. Further testing will reveal if I can avoid
that as slot reuse is preferable to lots of boilerplate attributes,
which are largely identical for each class.