Resolve coercion issue of LinkML data classes
This appears to have arisen from lack of the `inlined' and
`inlined_as_list' directives in the `sources' section of the `JobDesc'
class.

The result was that this list element would be coerced to a list and/or
dictionary, depending on usage and conditions. (I was able to trigger
the latter in the Python shell when I was exploring the API separately
from `fisdat'.)

The distinction between a multi-valued section with and without these
directives was pretty subtle, and was only be triggered if `identifier'
was toggled in one of the attributes/slots. As such, this may be a bug in
both the LinkML calls to RDFLib and the documentation.

Since `identifier' is now set for all of the classes in the `job.yaml'
section of our schema/data model (`TableDesc', `SourceDesc', `JobDesc'
and ultimate `ManifestDesc'), each on the `atomic_name' attribute/slot,
the `atomic_name' portion goes away, which makes the resulting turtle
files easier to read.