Add helper function for malformed IDs
This seems to be an issue when converting back and forth between the
LinkML YAML representation and turtle.

For the `identifier' attribute, the conversion to YAML still preserves
the `prefix:atom' format, but will accept both this and the atom alone.

Unfortunately, while it will error for duplicate identifiers which are
exactly right, it does not throw an error when one is prefixed with the
prefix, and one without. The consequence of which is the higher set of
triples in the file will be preserved, and the lower set of triples will
silently fail to be added.