Use XSD links for as many fields as possible
These are fundamentally links to objects in other schema, although I'm
not certain that this is the best way in which to consistently link to
these fields.

Namely, this is because they are links rather than elements which can be
used to form triples, and rdflib won't serialise unused prefixes.

The boilerplate moved into the utils script only applies to the example
job generation, but this should be helpful if we want to auto-generate
more of these. This works as follows:

1. Given a set of column slots in a given `TableSchema' instance,
   annotate the slots with the type (descriptive, collected or modelled).

   This is derived from the `is_a' slot attribute, which is defined in
   the main data model

2. Fetch the `exact_mappings', this is considered to be the underlying
   variable.