Serialise YAML by default, but upload TTL + YAML
The previous workflow involved:
1. Run `fisdat' a couple of times to append to a turtle manifest
2. To add jobs, run `fisjob' to convert to YAML, edit YAML, then run it again to generate turtle manifest
3. Run `fisup' to upload turtle manifest

This is annoying and it is fairly easy to overwrite a WIP manifest file.

This commit changes the workflow as follows:
1. Run `fisdat' a couple of times, which works on a YAML manifest file
2. Run `fisup' once, which makes the conversion to turtle of both the manifest file and any child schemata.

This is much easier to explain and there's only room for error once.
The functions do have a `--force' flag which we can switch on later,
if overwriting the wrong files becomes a problem.

Closely related to now converting the schemata to turtle, is that being
able to work directly on the manifest in `fisup' when converting to YAML
allows us to add the `schema_path_ttl' files to it before serialising it.
Then, there's a single upload command which pulls in everything.
In general, then, the internal representation is much better and the
behaviour of `fisup' is now much cleaner.

I have further removed the scoped columns flag from the `fisdat' CLI.
This was used for generating examples, whereas I removed the
example-generating code earlier, so it is defunct, not to mention
confusing.