Pin LinkML Python versions to avoid breaking changes
The exact relation between the two LinkML libraries (`linkml' and
`linkml-runtime') is somewhat vague (at least, in terms of the version
numbers, which are similar but out of sync), and it was in linkml-runtime
v1.7.6 in which a breaking change was introduced. Pin these for now.

There is an argument to pin the `google-cloud-storage' library, but its
API is less likely to change, and we use very little of it, so leave it
alone for now.

The `versioneer' library is partly 'vendored' and will probably be
removed if we can't get it working with older Python versions where
Python version is a 'hard' requirement such as on certain corporate
networks/laptops.