Rudimentary unit testing, at least for manifest initialisation
The program operation is now sufficiently complicated that testing this
manually was annoying. There are several modes, and two formats on which
to operate (YAML and TTL), which behave slighly differently.

This can be run using `python -m unittest fisdat.test_cmd_dat.TestInit`

This highlighted behaviour where the *manifest* is output in the current
working directory (even though the message printed says otherwise).
The behaviour we want is for it to be output anywhere (in the tests, this
goes in `/tmp/`), only that the data files are the file name alone.

Further to this, the main `fisdat` program should now print a warning if
a specified data file is not in the current working directory.