Further develop tests for the fisdat CLI program
This further tests the effect of overriding the base prefix (we want to
do enable this) and changing the data model URI.

This highlighted a couple of things.

First, there's a slight advantage to being able to change the
data model/ontology URI (which points to where some `meta.yaml' file
lives) particularly for testing, like here. In contrast, because the
`saved' prefix is hardcoded into almost all files already, there's no
advantage to being able to change that.

Second, in the former case, this is a network resource which needs to be
fetched, so we need to catch any errors anyway, and it's feasible that we
point it to some more specific URI like some 'job.yaml' for testing.
In the latter case, the saved: prefix is hardcoded into
`fisdat.data_model' (file `fisdat/data_model.py'), and it should never
change.

The actual advantage is pretty slight, however, and there's a good
argument to not be able to change the data model URI either, at least as
a command line option.

For now, then, remove the option to change the `saved' prefix from the
command line tools. This is just confusing.