Add additional test cases for appending unvalidated data
Although the schema in these tests is not matched to the data file,
these tests should succeed because validation is disabled.

Testing this further highlighted that my check for the success/failure
followed by removing the successfully-created file, was not working.
We try removing the file then assert false the conversion of the error to
true/false (which indicates whether the error comes through).
This is a bit of a hack but it makes the tests fail successfully.