Finish the tests of validation, minus the --dry-run option.
At the moment, the test-generating function is still pretty rough,
and using it is fairly confusing.

When I was writing these tests initially, I stated that there shouldn't
be an option to enable/disable validation of the schemata at all.
In fact, the validation is implicit in that we attempt to convert from
the YAML schema format to TTL. So, the assumption is that these are valid
YAML files. Thus, enabling/disabling the conversion is something of a
validation step, so test enabling/disabling this.

I further encountered problems with effects, where updating the object
passed as an argument would update the object outside of the function,
when I supposed that returning it would be a new object. What worked to
avoid this was to make a deep copy of the loaded manifest object, then
pass its tables into the table processing function. What didn't work was
making a copy of that list with the list's `.copy()' method - it would
still update the parent manifest object!