Flesh out the remaining tests for the --dry-run option
This means that we're able to test that the thing didn't actually write
to the files in question (they're still empty).

Further highlights that we should do the assertion before removing the
files, then we can assert both for the presence/absence of files.
(Change the tests for the `fisdat' utility to behave like this, too.)