Tweak error messages and feedback for CLI usage
On their own, these changes appear relatively minor, but they are largely
the implementation of the tests, which really stress the various function
cases consistently and robustly.

The warning/error messages now point to the command line options (e.g.
when to use the `--force'/`-F' option), and a related change is to
include ".converted" in the converted file name extensions.

Implementing the tests seems to have fixed an issue I was having with the
`--dry-run' option, where it would print the TTL manifest name twice in
the generated 'index' file, which is what I was hoping.