Annotate any manifest originally generated as TTL
This is necessary because the converted TTL schema file name was only
being added to the *converted* manifest file. Given an RDF/TTL input
manifest, then, only the converted YAML manifest file had this
information appended.

This means that the pipeline, which does not read the YAML schema files,
reads a TTL manifest without the TTL schema file name, and its validation
of the RDF graph fails as expected, since this field is nil.

Therefore, when the manifest file provided to fisup(1) is RDF/TTL,
generate an additional RDF/TTL manifest file with this information
added, notionally this is the equivalent of the input manifest, just
annotated with this field.

This may come in useful later if there are any additional fields which
need to be added to the manifest file prior to upload.

Unit tests seem to pass. The rather tortured test generating function for
testing various manifest cases needed a bit of work, but it highlighted
the behaviour of the --dry-run CLI option where the annotated manifest
wasn't being returned. There will be some manual testing to do.