Always forcibly overwrite schema conversions
It turned out that when data files share a schema, the program tries to
convert these again, and it fails if conversion had already taken place.

Fixing this robustly really requires a change to the data model (i.e.,
declare schema files as a resource, and only make a conversion once per
resource), so always overwrite for now.

There isn't any real problem in making this change because these are not
written by hand. (The conversion is an RDF equivalent of the file and
uses a bunch of stuff from various ontologies). So, there should not be
any work lost when overwriting these. Further, they have '.converted'
added to the file extension so there's no chance that a file would be
overwritten which wasn't this target conversion.

I have additionally removed the unit tests which tested the conversion of
schema files. These are commented out and can be added back in if and
when we fix the data model.