Enforce Python version dependency
The linkml/linkml-runtime packages appear to need at least Python 3.8,
so enforce Python version requirement.

This is further to the issue with versioneer for semantic versioning,
which was fragile, and could be removed, but removing this is not
possible, and we'd need to change how we write data schemata if we used
something else.