Fetch version string from `setup.py`
I hadn't removed everything related to the broken versioneer, and we
already depend on setuptools, so just use that to get what we specify in
`setup.py`.

This will probably require remaking the virtual environment as it needed
to be added as a dependency there (not sure why, since it's a dependency
when creating the virtual environment and is imported by `setup.py`).